third_party/blink/web_tests/external/wpt/css/css-flexbox/flex-container-max-content-002.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] .flex 5
  assert_equals: \n<div class="flex min" style="flex-flow: row nowrap" data-expected-width="180" data-expected-height="45">\n  <div data-expected-width="85" data-expected-height="35">X X</div>\n  <div data-expected-width="85" data-expected-height="35">X X</div>\n</div>\nheight expected 45 but got 20
[FAIL] .flex 6
  assert_equals: \n<div class="flex max" style="flex-flow: row nowrap" data-expected-width="180" data-expected-height="45">\n  <div data-expected-width="85" data-expected-height="35">X X</div>\n  <div data-expected-width="85" data-expected-height="35">X X</div>\n</div>\nheight expected 45 but got 210
Harness: the test ran to completion.
